requirements.txt
streamlit
google-genai
python-dotenv
pandas
psycopg2-binary
sqlalchemy
PyPDF2